devop: move home layout under core
