+++ src/core/viewDefault/components/homeLayout/HomeFeatures.vue
