+++ src/core/viewDefault/layouts/TheHomeLayout.vue
@@ -1,8 +1,8 @@
 <template>
   <div>
-    <landing />
-    <features />
-    <why-mew />
+    <home-landing />
+    <home-features />
+    <home-why-mew />
     <join />
     <partners />
     <get-started />
@@ -10,19 +10,19 @@
 </template>
 
 <script>
-import landing from '../../components/landing/Landing';
-import features from '../../components/features/Features';
-import whyMew from '../../components/why-mew/WhyMew';
+import HomeLanding from '../components/homeLayout/HomeLanding';
+import HomeFeatures from '../components/homeLayout/HomeFeatures';
+import HomeWhyMew from '../components/homeLayout/HomeWhyMew';
 import join from '@/components/join/Join';
 import getStarted from '@/components/get-started/GetStarted';
 import partners from '@/components/partners/Partners';
 
 export default {
-  name: 'LandingPage',
+  name: 'TheHomeLayout',
   components: {
-    landing,
-    features,
-    whyMew,
+    HomeLanding,
+    HomeFeatures,
+    HomeWhyMew,
     join,
     getStarted,
     partners
